Added load generator harness; plus significant changes resulting from moving the syncing of the local and distributed log out of kernel.

+++ src/dl_produce_request.h
@@ -72,9 +72,9 @@ struct dl_produce_request {
 };
 
 extern int dl_produce_request_new(struct dl_request **, const int32_t,
-    struct sbuf *, int16_t, int16_t, struct sbuf *, struct dl_message_set *);
+    struct sbuf *, int16_t, int32_t, struct sbuf *, struct dl_message_set *);
 extern int dl_produce_request_new_nomsg(struct dl_request **,
-    const int32_t, struct sbuf *, int16_t, int16_t, struct sbuf *);
+    const int32_t, struct sbuf *, int16_t, int32_t, struct sbuf *);
 extern void dl_produce_request_delete(struct dl_produce_request *);
 
 extern int dl_produce_request_decode(struct dl_produce_request **,
